% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotateVec}
\alias{annotateVec}
\title{annotate vector of petpide sequences against fasta file}
\usage{
annotateVec(pepseq, fasta, digestPattern = "(([RK])|(^))", mcCores = NULL)
}
\arguments{
\item{pepseq}{peptide sequences}

\item{fasta}{fasta file}

\item{digestPattern}{digest pattern as regex}

\item{mcCores}{nr of cores to use}
}
\description{
annotate vector of petpide sequences against fasta file
}
\examples{

library(prozor)
file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = readPeptideFasta(file = file)

res = annotateVec(pepdata[1:20,"peptideSequence"],fasta)
head(res)
}

