% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotateVec2}
\alias{annotateVec2}
\title{annotate peptides using AhoCorasickTrie}
\usage{
annotateVec2(pepinfo, fasta, digestPattern = c("", "K", "R"))
}
\arguments{
\item{pepinfo}{- list of peptides - sequence, optional modified sequence, charge state.}

\item{fasta}{- object as created by read.fasta in pacakge seqinr}

\item{digestPattern}{- default "(([RK])|(^)|(^M))"}

\item{mcCores}{number of cores to use}
}
\description{
peptides which do not have protein assignment drop out
}
\examples{

library(AhoCorasickTrie)

file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = read.fasta(file = file, as.string = TRUE, seqtype="AA")
res = annotatePeptides(pepdata[1:20,], fasta,mcCores=1)
res = annotatePeptides2(pepdata[1:20,"peptideSequence"],fasta)
head(res)


}

