% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz_helper.R
\name{bfabric_save_fasta}
\alias{bfabric_save_fasta}
\title{copy to server}
\usage{
bfabric_save_fasta(
  project_id,
  filepath,
  summary,
  python_path = "c:/Program Files/Python38"
)
}
\arguments{
\item{project_id}{project_id}

\item{filepath}{filepath}

\item{summary}{summary (string with info about DB)}

\item{python_path}{python_path}
}
\description{
copy to server
}
\keyword{internal}
