% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fgcz_fasta_db.R
\name{create_fgcz_fasta_db}
\alias{create_fgcz_fasta_db}
\title{create fasta db from one or more fasta files}
\usage{
create_fgcz_fasta_db(
  databasedirectory,
  useContaminants = loadContaminantsFGCZ2022(),
  revLab = "REV_",
  outputdir = NULL
)
}
\arguments{
\item{databasedirectory}{directory with fasta files}

\item{useContaminants}{contaminants to add}

\item{revLab}{reverse label}

\item{outputdir}{output directory}
}
\value{
list list(resDB, filepath , summary, mcall, dbname)
}
\description{
create fasta db from one or more fasta files
}
\examples{
print("NO exmple since function also writes the fasta files")

}
