% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm}
\usage{
greedy(pepprot)
}
\arguments{
\item{pepprot}{matrix as returned by prepareMatrix}
}
\value{
list of protein peptide assignemnts (where winner takes all)
}
\description{
given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm
}
\examples{
library(prozor)

data(protpepmetashort)
xx = prepareMatrix(protpepmetashort, weight= "count")
dim(xx)
es = greedy(as.matrix(xx))
}

