% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_parsimony}
\alias{greedy_parsimony}
\title{given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm}
\usage{
greedy_parsimony(pepprot)
}
\arguments{
\item{pepprot}{matrix as returned by prepareMatrix}
}
\value{
list of peptide protein assignment
}
\description{
given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm
}
\examples{
#library(prozor)

data(protpepmetashort)
colnames(protpepmetashort)
dim(unique(protpepmetashort[,4]))
xx = prepareMatrix(protpepmetashort, peptideID = "peptideModSeq")
dim(xx)
stopifnot(dim(xx)[1] == dim(unique(protpepmetashort[,4]))[1])

es = greedy(xx)
debug(prozor:::.greedy2)
stopifnot(length(unique(names(es))) == dim(unique(protpepmetashort[,4]))[1])

}
