% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedyRes2Matrix}
\alias{greedyRes2Matrix}
\title{converts result of greedy function to a matrix with 3 columns - peptide - charge and protein}
\usage{
greedyRes2Matrix(res)
}
\arguments{
\item{res}{result of function prozor::greedy}
}
\value{
matrix of peptide protein assignments
}
\description{
converts result of greedy function to a matrix with 3 columns - peptide - charge and protein
}
