% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsw.R
\name{hardconstrain}
\alias{hardconstrain}
\title{tests hard constraints}
\usage{
hardconstrain(splits, minwindow = 5, maxwindow = 50)
}
\arguments{
\item{splits}{the proposed splits}

\item{minwindow}{min window size}

\item{maxwindow}{max window size}
}
\value{
logical
}
\description{
tests hard constraints
}
\keyword{internal}
