% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{load_special_prot_2024}
\alias{load_special_prot_2024}
\title{load special prot}
\usage{
load_special_prot_2024(noHuman = FALSE)
}
\arguments{
\item{noHuman}{should human contaminants be excluded? default FALSE}
}
\value{
list with contaminant sequences
}
\description{
load special prot
}
\examples{
#library(prozor)
cont <- load_special_prot_2024()
length(cont)
contNH <- load_special_prot_2024(noHuman = TRUE)
length(contNH)
#example how to create a protein db with decoy sequences
}
