% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{load_universal_contaminants_2024}
\alias{load_universal_contaminants_2024}
\title{load universal contaminants}
\usage{
load_universal_contaminants_2024(noHuman = FALSE)
}
\arguments{
\item{noHuman}{should human contaminants be excluded? default FALSE}
}
\value{
list with contaminant sequences
}
\description{
These sequences are downloaded from here https://github.com/HaoGroup-ProtContLib/Protein-Contaminant-Libraries-for-DDA-and-DIA-Proteomics
}
\examples{
#library(prozor)
cont <- load_universal_contaminants_2024()
length(cont)
contNH <- load_universal_contaminants_2024(noHuman = TRUE)
length(contNH)
#example how to create a protein db with decoy sequences
}
