% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsFasta}
\alias{loadContaminantsFasta}
\title{load list of contaminant sequences}
\usage{
loadContaminantsFasta()
}
\description{
load list of contaminant sequences
}
\examples{
library(prozor)
cont <- loadContaminantsFasta()
cont[[1]]
#example how to create a protein db with decoy sequences
}
