% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsFasta2019}
\alias{loadContaminantsFasta2019}
\title{load list of contaminant sequences FGCZ 2019}
\usage{
loadContaminantsFasta2019()
}
\description{
load list of contaminant sequences FGCZ 2019
}
\examples{
library(prozor)
cont <- loadContaminantsFasta2019()
cont[[1]]
#example how to create a protein db with decoy sequences
}
