% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsFasta2021}
\alias{loadContaminantsFasta2021}
\title{load list of contaminant sequences FGCZ 2021}
\usage{
loadContaminantsFasta2021(noHuman = FALSE)
}
\arguments{
\item{noHuman}{should human contaminants be excluded? default FALSE}
}
\description{
load list of contaminant sequences FGCZ 2021
}
\examples{
library(prozor)
cont <- loadContaminantsFasta2021()
length(cont)
contNH <- loadContaminantsFasta2021()
length(contNH)
#example how to create a protein db with decoy sequences
}
