% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsFGCZ2022}
\alias{loadContaminantsFGCZ2022}
\title{load list of contaminant sequences FGCZ 2022}
\usage{
loadContaminantsFGCZ2022(noHuman = FALSE)
}
\arguments{
\item{noHuman}{should human contaminants be excluded? default FALSE}
}
\value{
list with contaminant sequences
}
\description{
load list of contaminant sequences FGCZ 2022
}
\examples{
#library(prozor)
cont <- loadContaminantsFGCZ2022()
length(cont)
contNH <- loadContaminantsFGCZ2022(noHuman = TRUE)
length(contNH)
#example how to create a protein db with decoy sequences
}
