% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsNoHumanFasta}
\alias{loadContaminantsNoHumanFasta}
\title{load list of contaminant without human sequences}
\usage{
loadContaminantsNoHumanFasta()
}
\description{
load list of contaminant without human sequences
}
\examples{
library(prozor)
cont <- loadContaminantsNoHumanFasta()
cont[[1]]
#example how to create a protein db with decoy sequences
}
