% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadHomoSapiensFasta}
\alias{loadHomoSapiensFasta}
\title{load human}
\usage{
loadHomoSapiensFasta()
}
\description{
load human
}
\examples{
library(prozor)
fastaMusMusc <- loadHomoSapiensFasta()

}
