% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadHomoSapiensSignalPeptides}
\alias{loadHomoSapiensSignalPeptides}
\title{load human signal peptides}
\usage{
loadHomoSapiensSignalPeptides()
}
\description{
load human signal peptides
}
\examples{
library(prozor)
signal <- loadHomoSapiensSignalPeptides()

}
