% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadMusMusculusFasta}
\alias{loadMusMusculusFasta}
\title{load mus musculus fasta}
\usage{
loadMusMusculusFasta()
}
\description{
load mus musculus fasta
}
\examples{
library(prozor)
fastaMusMusc <- loadMusMusculusFasta()

}

