% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadMusMusculusSignalPeptides}
\alias{loadMusMusculusSignalPeptides}
\title{load mus musculus signal peptides}
\usage{
loadMusMusculusSignalPeptides()
}
\description{
load mus musculus signal peptides
}
\examples{
library(prozor)
signal <- loadMusMusculusSignalPeptides()
head(signal)
}
