% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fgcz_fasta_db.R
\name{make_fasta_summary}
\alias{make_fasta_summary}
\title{make db summary which includes number of sequences, and amino acid statistcis}
\usage{
make_fasta_summary(resDB, old = FALSE, as_string = TRUE)
}
\value{
array of strings which should be passed to the cat function.
}
\description{
make db summary which includes number of sequences, and amino acid statistcis
}
\examples{
file = system.file("extdata/fgcz_contaminants2022_20220405.fasta.gz",package="prozor")
xx <- prozor::readPeptideFasta(file)
cat(make_fasta_summary(xx))
rbenchmark::benchmark(make_fasta_summary(xx),replications = 500)
rbenchmark::benchmark(make_fasta_summary(xx, old = TRUE), replications = 500)
}
