% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSignalPeptides.R
\name{makeID}
\alias{makeID}
\title{make id for chain in format sp|P30443|1A01_HUMANs25}
\usage{
makeID(sequence, id, sp)
}
\arguments{
\item{sequence}{- aa sequence as string}

\item{id}{uniprot id id: sp|P30443|1A01_HUMAN}

\item{sp}{start position of chain numeric}
}
\description{
make id for chain in format sp|P30443|1A01_HUMANs25
}
\examples{
seq <- "MAVMAPRTLLLLLSGALALTQTWAGSHSMRYFFTSVSRPGR\\
GEPRFIAVGYVDDTQFVRFDSDAASQKMEPRAPWIEQEGPEYWDQETRN\\
MKAHSQTDRANLGTLRGYYNQSEDGSHTIQIMYGCDVGPDGRFLRGYRQ\\
DAYDGKDYIALNEDLRSWTAADMAAQITKRKWEAVHAAEQRRVYLEGRC\\
VDGLRRYLENGKETLQRTDPPKTHMTHHPISDHEATLRCWALGFYPAEI\\
TLTWQRDGEDQTQDTELVETRPAGDGTFQKWAAVVVPSGEEQRYTCHVQ\\
HEGLPKPLTLRWELSSQPTIPIVGIIAGLVLLGAVITGAVVAAVMWRRK\\
SSDRKGGSYTQAASSDSAQGSDVSLTACKV"
nam <-"sp|P30443|1A01_HUMAN"
sp <- 24
makeID(seq, nam, sp)
}
