% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSignalPeptides.R
\name{makeIDUnip}
\alias{makeIDUnip}
\title{make id for chain compatible with uniprot}
\usage{
makeIDUnip(sequence, id, sp)
}
\arguments{
\item{sequence}{- aa sequence as string}

\item{id}{uniprot id id: sp|P30443|1A01_HUMAN}

\item{sp}{start position of chain numeric}
}
\description{
make id for chain compatible with uniprot
}
\examples{
seq <- "MAVMAPRTLLLLLSGALALTQTWAGSHSMRYFFTSVSRPGR\\
GEPRFIAVGYVDDTQFVRFDSDAASQKMEPRAPWIEQEGPEYWDQETRN\\
MKAHSQTDRANLGTLRGYYNQSEDGSHTIQIMYGCDVGPDGRFLRGYRQ\\
DAYDGKDYIALNEDLRSWTAADMAAQITKRKWEAVHAAEQRRVYLEGRC\\
VDGLRRYLENGKETLQRTDPPKTHMTHHPISDHEATLRCWALGFYPAEI\\
TLTWQRDGEDQTQDTELVETRPAGDGTFQKWAAVVVPSGEEQRYTCHVQ\\
HEGLPKPLTLRWELSSQPTIPIVGIIAGLVLLGAVITGAVVAAVMWRRK\\
SSDRKGGSYTQAASSDSAQGSDVSLTACKV"
nam <-"sp|P30443|1A01_HUMAN"
sp <- 24
makeIDUnip(seq, nam, sp)
}
