% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsw.R
\docType{data}
\name{masses}
\alias{masses}
\title{MS masses
A dataset containing approx 150000 MS1 precursor masses}
\format{
An object of class \code{numeric} of length 146009.
}
\usage{
masses
}
\description{
MS masses
A dataset containing approx 150000 MS1 precursor masses
}
\keyword{datasets}
