% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsw.R
\name{objectiveMS1Function}
\alias{objectiveMS1Function}
\title{compute the deviation from optimum: equal number of MS1 per bin}
\usage{
objectiveMS1Function(splits, data)
}
\arguments{
\item{splits}{the new window boundaries}

\item{data}{the data}
}
\value{
list with score1 - manhattan distance, score2 - euclidean distance, counts - observed counts, optimumN - optimum counts
}
\description{
compute the deviation from optimum: equal number of MS1 per bin
}
