% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFDR.R
\name{predictScoreFDR}
\alias{predictScoreFDR}
\title{Predict score given FDR}
\usage{
predictScoreFDR(fdrObj, qValue = 1, method = "SimpleFDR")
}
\arguments{
\item{fdrObj}{object generated by computeFDR}

\item{qValue}{false discovery rate in percent, default 1 percent}

\item{method}{either FPR or SimpleFDR, default is SimpleFDR}
}
\value{
score for a given FDR
}
\description{
For more details and references see package vignette
\code{vignette("TargetDecoyFDR_Example", package = "prozor")}
}
\examples{
data(fdrSample)
fdr1 <- computeFDRwithID(fdrSample$score, fdrSample$proteinID, larger_better = FALSE)

predictScoreFDR(fdr1,qValue=5)
fdr2<-computeFDRwithID(fdrSample$score2, fdrSample$proteinID, larger_better = TRUE)
predictScoreFDR(fdr2,qValue=5)

}
