% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{given table of peptide protein assigments generate matrix}
\usage{
prepareMatrix(data, weight = "count")
}
\arguments{
\item{data}{generated by annotatePeptides}

\item{weight}{weight type to use options are "count", "AA" "coverage"}
}
\value{
sparse matrix
}
\description{
given table of peptide protein assigments generate matrix
}
\examples{
data(protpepmetashort)
library(Matrix)
colnames(protpepmetashort)
dim(protpepmetashort)
xx = prepareMatrix(protpepmetashort, weight= "count")
#xx = prepareMatrix(protpepmetashort, weight = "AA")
#xx = prepareMatrix(protpepmetashort, weight = "coverage")
image(as.matrix(xx))
corProt = cor(as.matrix(xx))
par(mfrow =c(2,1))
image(corProt)
rownames(corProt)

# penalise peptides matching many proteins
xxn = t(apply(xx, 1, function(x){x/sum(x)}))
dim(xxn)
corProtn = cor(as.matrix(xxn))
image(corProtn)
rownames(corProtn)
}

