% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{protpepmetashort}
\alias{protpepmetashort}
\title{Small version of prot meta dataset to speed up computation}
\description{
Small version of prot meta dataset to speed up computation
}
\keyword{data}

