% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSignalPeptides.R
\name{removeSignalPeptide}
\alias{removeSignalPeptide}
\title{remove signal peptides from main chain}
\usage{
removeSignalPeptide(db, signal, idfun = makeID)
}
\arguments{
\item{db}{uniprot fasta database as list}

\item{signal}{tab delimited file with signals}

\item{idfun}{function to generate id's}
}
\description{
remove signal peptides from main chain
}
\examples{
\dontrun{
library(prozor)

hsfasta <- loadHomoSapiensFasta()
hssignal <- loadHomoSapiensSignalPeptides()
xx <- removeSignalPeptide(hsfasta, hssignal)
db <- hsfasta
signal <- hssignal
}
}
