% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcz_helper.R
\name{scp_copy}
\alias{scp_copy}
\title{copy to server}
\usage{
scp_copy(
  filepath,
  to = "/srv/www/htdocs/FASTA",
  username = "wolski",
  host = "fgcz-r-021.uzh.ch"
)
}
\arguments{
\item{filepath}{path of file}

\item{to}{destination folder}

\item{username}{username}

\item{host}{host}
}
\value{
copies file using scp
}
\description{
copy to server
}
\keyword{internal}
