% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotateAHO}
\alias{annotateAHO}
\title{annotate peptides using AhoCorasickTrie}
\usage{
annotateAHO(pepseq, fasta)
}
\arguments{
\item{pepseq}{- list of peptides - sequence, optional modified sequence, charge state.}

\item{fasta}{- object as created by \code{readPeptideFasta}}
}
\value{
A data.frame with proteinID, peptideSeq, Offset and proteinSequence
}
\description{
peptides which do not have protein assignment drop out
}
\examples{

#library(prozor)
library(AhoCorasickTrie)
file = system.file("p1000_db1_example/shortfasta.fasta.gz",package = "prozor")
fasta = readPeptideFasta(file = file)
pepprot <- get(data("pepprot", package = "prozor"))
system.time( res2 <- annotateAHO( pepprot[seq_len(20),"peptideSeq"], fasta))
colnames(res2)
class(res2)
}
