% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotatePeptides}
\alias{annotatePeptides}
\title{Annotate peptides with protein ids}
\usage{
annotatePeptides(pepinfo, fasta, digestPattern = "(([RK])|(^))",
  mcCores = NULL)
}
\arguments{
\item{pepinfo}{- list of peptides - sequence, optional modified sequence, charge state.}

\item{fasta}{- object as created by readPeptideFasta}

\item{digestPattern}{- default "(([RK])|(^))"}

\item{mcCores}{number of cores to use}
}
\description{
peptides which do not have protein assignment drop out
}
\examples{
library(prozor)
library(doParallel)
library(foreach)
data(pepdata)
head(pepdata)

file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = readPeptideFasta(file = file)
res = annotatePeptides(pepdata[1:20,], fasta,mcCores=1)
res = annotatePeptides(pepdata[1:20,"peptideSequence"],fasta)
head(res)

}

