% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotatePeptides}
\alias{annotatePeptides}
\title{Annotate peptides with protein ids}
\usage{
annotatePeptides(pepinfo, fasta, digestPattern = c("R", "K", ""))
}
\arguments{
\item{pepinfo}{- list of peptides - sequence, optional modified sequence, charge state.}

\item{fasta}{- object as created by readPeptideFasta}

\item{digestPattern}{- default "(([RK])|(^))"}
}
\description{
peptides which do not have protein assignment drop out
}
\examples{
library(prozor)
data(pepdata)
head(pepdata)

file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = readPeptideFasta(file = file)
res = annotatePeptides(pepdata[1:20,], fasta)
head(res)
res = annotatePeptides(pepdata[1:20,"peptideSequence"],fasta)
length(res)
names(res[[1]])
names(res[[2]])
head(res[[2]])
class(res[[1]])
class(res[[2]])
head(res[[1]])
x<-merge(res[[1]],data.frame(res[[2]]))
res = merge(res[[1]],res[[2]],by.x="peptideSequence",by.y="peptideSequence")
head(res)

}

