% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotateVec}
\alias{annotateVec}
\title{annotate vector of petpide sequences against fasta file}
\usage{
annotateVec(pepseq, fasta)
}
\arguments{
\item{pepseq}{peptide sequences}

\item{fasta}{fasta file}
}
\description{
annotate vector of petpide sequences against fasta file
}
\examples{

library(prozor)
file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = read.fasta(file = file, as.string = TRUE, seqtype="AA")

res = annotateVec(pepdata[1:20,"peptideSequence"],fasta)
}

