% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{annotateVec2}
\alias{annotateVec2}
\title{annotate peptides using AhoCorasickTrie}
\usage{
annotateVec2(pepseq, fasta, digestPattern = c("", "K", "R"))
}
\arguments{
\item{pepseq}{- list of peptides - sequence, optional modified sequence, charge state.}

\item{fasta}{- object as created by readPeptideFasta}

\item{digestPattern}{- list of N terminal amino acids including empty string (protein start)}
}
\description{
peptides which do not have protein assignment drop out
}
\examples{


file = file.path(path.package("prozor"),"extdata/shortfasta.fasta" )
fasta = readPeptideFasta(file = file)
res = annotateVec(pepdata[1:20,"peptideSequence"],fasta)
head(res)
res2 = annotateVec2(pepdata[1:20,"peptideSequence"],fasta)
head(res2)
colnames(res2)
}

