% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsw.R
\docType{class}
\name{Cdsw-class}
\alias{Cdsw-class}
\alias{Cdsw}
\title{Compute dynamic swath windows}
\description{
Compute dynamic swath windows
}
\section{Fields}{

\describe{
\item{\code{masses}}{MS1 masses}

\item{\code{breaks}}{the breaks}

\item{\code{nbins}}{number of bins}

\item{\code{digits}}{mass accuracy in result}
}}

\section{Methods}{

\describe{
\item{\code{asTable(overlap = 1)}}{make windows}

\item{\code{error()}}{show error}

\item{\code{optimizeWindows(digits = 1, maxbin = 15, plot = FALSE, overlap = 0)}}{optimizes the windows}

\item{\code{quantile_breaks(digits = 2)}}{same number of MS1 in each window but might violate hard constraints}

\item{\code{sampling_breaks(maxwindow = 150, minwindow = 5, digits = 2, plot = FALSE)}}{starts with quantile breaks but mixes with uniform data to satisfy had constraints}

\item{\code{showCycle(overlap = 1)}}{shows the generated DIA cycle}
}}

\examples{
library(cdsw)
cdsw <- Cdsw(masses)
cdsw$sampling_breaks(maxwindow=100,plot=TRUE)
cdsw$plot()
cdsw$asTable()
cdsw$breaks
}
