% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFDR.R
\name{computeFDR}
\alias{computeFDR}
\title{compute FDR given a score}
\usage{
computeFDR(score, proteinID, decoy = "REV_", decreasing = TRUE)
}
\arguments{
\item{score}{a vector with scores}

\item{proteinID}{- list with protein id's}

\item{decoy}{decoy pattern, default "REV_"}

\item{decreasing}{If the scores sorted from small (better),
than decreasing=FALSE, or if from large (better),
than decreasing TRUE (default).}
}
\value{
list with proteinID, decoy_hit (indicates if decoy), score the search engine score,
FDR1 false discovery rate estimated using the method of Gyggi, FDR2 - estimated using the method of Kell.
}
\description{
compute FDR given a score
}

