% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fgcz_fasta_db.R
\name{create_fgcz_fasta_db}
\alias{create_fgcz_fasta_db}
\title{create fasta db from one or more fasta files}
\usage{
create_fgcz_fasta_db(
  databasedirectory,
  useContaminants = loadContaminantsFasta(),
  revLab = "REV_"
)
}
\description{
create fasta db from one or more fasta files
}
