% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{filterSequences}
\alias{filterSequences}
\title{Filter for specific residues}
\usage{
filterSequences(matches, digestPattern = c("", "K", "R"))
}
\arguments{
\item{matches}{must have 2 columns proteinSequnce and Offset}

\item{digestPattern}{- list of N terminal amino acids including empty string (protein start) default tryptic = c("","K","R")}
}
\description{
Will check if AA at Offset is a valid cleavage site
}

