% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeptides.R
\name{filterSequences}
\alias{filterSequences}
\title{Filter for specific residues}
\usage{
filterSequences(matches, prefix = "(([RK])|(^)|(^M))", suffix = "")
}
\arguments{
\item{matches}{must have 2 columns proteinSequnce and Offset}

\item{prefix}{- regular expression describing the prefix of the peptide sequence e.g. (([RK])|(^)|(^M))}

\item{suffix}{- regular expression describing the suffix of the peptide sequence}
}
\description{
Will check if AA at Offset is a valid cleavage site
}
