% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy}
\alias{greedy}
\title{given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm}
\usage{
greedy(pepprot)
}
\arguments{
\item{pepprot}{matrix as returned by prepareMatrix}
}
\value{
list of peptide protein assignment
}
\description{
given matrix (columns protein rows peptides), compute minimal protein set using greedy algorithm
}
\examples{
library(prozor)

data(protpepmetashort)
colnames(protpepmetashort)
dim(unique(protpepmetashort[,4:5]))
xx = prepareMatrix(protpepmetashort, weight= "one")
dim(xx)
stopifnot(dim(xx)[1] == dim(unique(protpepmetashort[,4:5]))[1])
es = greedy(as.matrix(xx))
stopifnot(length(unique(names(es))) == dim(unique(protpepmetashort[,4:5]))[1])

}
