% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContaminantsFasta.R
\name{loadContaminantsFasta2019}
\alias{loadContaminantsFasta2019}
\title{load list of contaminant sequences FGCZ 2019}
\usage{
loadContaminantsFasta2019(noHuman = FALSE)
}
\arguments{
\item{noHuman}{should human contaminants be excluded? default FALSE}
}
\value{
list with contaminant sequences
}
\description{
load list of contaminant sequences FGCZ 2019
}
\examples{
#library(prozor)
cont <- loadContaminantsFasta2019()
length(cont)
contNH <- loadContaminantsFasta2019()
length(contNH)
#example how to create a protein db with decoy sequences
}
