% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{masses}
\alias{masses}
\title{MS masses
A dataset containing approx 150000 MS1 precursor masses}
\description{
MS masses
A dataset containing approx 150000 MS1 precursor masses
}
\keyword{data}
