% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/greedy.R
\name{occam}
\alias{occam}
\title{given matrix}
\usage{
occam(pepprot)
}
\arguments{
\item{pepprot}{matrix as returned by prepareMatrix}
}
\description{
given matrix
}
\examples{
library(prozor)

data(protpepmeta)
xx = prepareMatrix(protpepmeta, weight= "count")
dim(xx)
es = occam(as.matrix(xx))
}

