% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFDR.R
\name{plotFDR}
\alias{plotFDR}
\title{plot FDR}
\usage{
plotFDR(data)
}
\arguments{
\item{data}{data returned by computeFDR function}
}
\description{
For more details and references see package vignette
\code{vignette("TargetDecoyFDR_Example", package = "prozor")}
}
\examples{
library(prozor)
data(fdrSample)
fdr1 <- computeFDRwithID(fdrSample$score, fdrSample$proteinID, larger_better = FALSE)
fdr2 <- computeFDRwithID(fdrSample$score2, fdrSample$proteinID, larger_better = TRUE)
plotFDR(fdr1)
plotFDR(fdr2)
data<-fdr1
}
