% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{given table of peptide protein assigments generate matrix}
\usage{
prepareMatrix(data, weight = "count")
}
\arguments{
\item{data}{generated by annotatePeptides}

\item{weight}{weight type to use options are "count", "AA" "coverage"}
}
\value{
sparse matrix
}
\description{
given table of peptide protein assigments generate matrix
}
\examples{
data(protpepmeta)
library(Matrix)
colnames(protpepmeta)
dim(protpepmeta)
xx = prepareMatrix(protpepmeta, weight= "count")
#xx = prepareMatrix(protpepmeta, weight = "AA")
#xx = prepareMatrix(protpepmeta, weight = "coverage")
image(as.matrix(xx[1:300,1:300]))
tmp = cor(as.matrix(xx))
par(mfrow =c(2,1))
image(tmp[1:300,1:300])
rownames(tmp[1:300,1:300])
dim(xx)
# penalise peptides matching many proteins
xxn = t(apply(xx, 1, function(x){x/sum(x)}))
dim(xxn)
tmpn = cor(as.matrix(xxn))
image(tmpn[1:300,1:300])
rownames(tmpn[1:300,1:300])
}

