% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{protpepmetashort}
\alias{protpepmetashort}
\title{Small version of pepprot dataset to speed up computation}
\description{
Small version of pepprot dataset to speed up computation
}
\keyword{data}
