% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFasta.R
\name{readPeptideFasta}
\alias{readPeptideFasta}
\title{wrapper setting the correct parameters
seqinr::read.fasta for reading peptide sequences}
\usage{
readPeptideFasta(file)
}
\arguments{
\item{file}{- fasta file}
}
\value{
list with sequences
}
\description{
peptides which do not have protein assignment drop out
}
\examples{
library(seqinr)

file = system.file("extdata/fgcz_contaminants2021_20210929.fasta.gz",package = "prozor")
fasta = readPeptideFasta(file)

}
