% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFasta.R
\name{readPeptideFasta}
\alias{readPeptideFasta}
\title{wrapper setting the correct parameters}
\usage{
readPeptideFasta(file)
}
\arguments{
\item{file}{- fasta file}
}
\description{
peptides which do not have protein assignment drop out
}
\examples{
library(seqinr)
library(prozor)
file = file.path(path.package("prozor"),"extdata/fgcz_contaminants_20150123.fasta")
fasta = readPeptideFasta(file)

}
